+++ msg/DetectionStatus.msg
@@ -1,4 +1,7 @@
 # This message is used for determining the status of the detection API results
+#  Raymond Kirk (Tunstill) Copyright (c) 2020
+#  Email: ray.tunstill@gmail.com
+
 bool OKAY # true if valid detections are returned
 bool BUSY # true if the detector is busy with another task
 bool ERROR # true if an error occurred
