+++ src/rasberry_perception/detection/interfaces/registry.py
